+++ Presentation.Web/Controllers/API/V1/DataProtectionAdvisorController.cs
@@ -13,7 +13,7 @@ using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API.V1
 {
-    [PublicApi]
+    [InternalApi]
     public class DataProtectionAdvisorController : GenericApiController<DataProtectionAdvisor, DataProtectionAdvisorDTO>
     {
         private readonly IGenericRepository<DataProtectionAdvisor> _repository;
