+++ Presentation.Web/Controllers/API/V1/ItContractController.cs
@@ -24,7 +24,7 @@ using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API.V1
 {
-    [PublicApi]
+    [InternalApi]
     [RoutePrefix("api/itcontract")]
     public class ItContractController : GenericHierarchyApiController<ItContract, ItContractDTO>
     {
