+++ Presentation.Web/Controllers/API/V1/ItSystemUsageOrgUnitUsageController.cs
@@ -11,7 +11,7 @@ using Presentation.Web.Models.API.V1;
 using Swashbuckle.Swagger.Annotations;
 namespace Presentation.Web.Controllers.API.V1
 {
-    [PublicApi]
+    [InternalApi]
     public class ItSystemUsageOrgUnitUsageController : BaseApiController
     {
         private readonly IGenericRepository<ItSystemUsageOrgUnitUsage> _responsibleOrgUnitRepository;
