+++ Presentation.Web/Controllers/API/V1/OData/ItContractAgreementElementTypesController.cs
@@ -11,7 +11,7 @@ namespace Presentation.Web.Controllers.API.V1.OData
     /// Gives access to relations between ItContract and ElementTypes
     /// Primarily used for reporting
     /// </summary>
-    [PublicApi]
+    [InternalApi]
     public class ItContractAgreementElementTypesController : BaseController<ItContractAgreementElementTypes>
     {
         public ItContractAgreementElementTypesController(IGenericRepository<ItContractAgreementElementTypes> repository)
