+++ Presentation.Web/Controllers/API/V1/OData/ItInterfacesController.cs
@@ -15,7 +15,7 @@ using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API.V1.OData
 {
-    [PublicApi]
+    [InternalApi]
     public class ItInterfacesController : BaseEntityController<ItInterface>
     {
         public ItInterfacesController(IGenericRepository<ItInterface> repository)
