+++ Presentation.Web/Controllers/API/V1/OData/LocalOptionControllers/LocalDataProcessingDataResponsibleOptionsController.cs
@@ -8,7 +8,7 @@ using Presentation.Web.Infrastructure.Attributes;
 
 namespace Presentation.Web.Controllers.API.V1.OData.LocalOptionControllers
 {
-    [PublicApi]
+    [InternalApi]
     [ODataRoutePrefix("LocalDataProcessingDataResponsibleOptions")]
     public class LocalDataProcessingDataResponsibleOptionsController : LocalOptionBaseController<LocalDataProcessingDataResponsibleOption, DataProcessingRegistration, DataProcessingDataResponsibleOption>
     {
