+++ Presentation.Web/Controllers/API/V1/PasswordResetRequestController.cs
@@ -7,14 +7,13 @@ using Core.DomainModel;
 using Core.DomainServices;
 using Infrastructure.Services.Cryptography;
 using Presentation.Web.Infrastructure.Attributes;
-using Presentation.Web.Models;
 using Presentation.Web.Models.API.V1;
 using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API.V1
 {
     [AllowAnonymous]
-    [PublicApi]
+    [InternalApi]
     [AllowRightsHoldersAccess]
     public class PasswordResetRequestController : BaseApiController
     {
