+++ Presentation.Web/Controllers/API/V1/TaskRefController.cs
@@ -5,7 +5,7 @@ using Presentation.Web.Models.API.V1;
 
 namespace Presentation.Web.Controllers.API.V1
 {
-    [PublicApi]
+    [InternalApi]
     public class TaskRefController : GenericHierarchyApiController<TaskRef, TaskRefDTO>
     {
         public TaskRefController(IGenericRepository<TaskRef> repository)
